+++ packages/workspace-fusion/src/lib/components/Workspace.tsx
@@ -5,16 +5,18 @@ import { WorkspaceConfiguration, WorkspaceProps } from '../types';
 import { didOptionsChange } from '../utils/optionsChanged/didOptionsChange';
 import { createConfigurationObject } from '../utils/createWorkspaceConfig';
 import { createWorkspaceController } from '../utils/createWorkspaceController';
+import { BaseEvent } from '@equinor/workspace-core';
 
 export function Workspace<
 	TData extends Record<PropertyKey, unknown>,
 	TContext extends Record<PropertyKey, unknown> = never,
+	TCustomSidesheetEvents extends BaseEvent = never,
 	TExtendedFields extends string = never,
 	TCustomGroupByKeys extends Record<PropertyKey, unknown> = never
->(props: WorkspaceProps<TData, TContext, TExtendedFields, TCustomGroupByKeys>) {
-	const configuration = useRef<WorkspaceConfiguration<TData, TContext, TExtendedFields, TCustomGroupByKeys>>(
-		createConfigurationObject(props)
-	);
+>(props: WorkspaceProps<TData, TContext, TCustomSidesheetEvents, TExtendedFields, TCustomGroupByKeys>) {
+	const configuration = useRef<
+		WorkspaceConfiguration<TData, TContext, TCustomSidesheetEvents, TExtendedFields, TCustomGroupByKeys>
+	>(createConfigurationObject(props));
 
 	/**
 	 * Triggers on every parent render and updates configuration accordingly