+++ packages/workspace-fusion/src/lib/components/provider/MediatorProvider.tsx
@@ -1,22 +1,25 @@
+import { BaseEvent } from '@equinor/workspace-core';
 import { createContext, ReactNode } from 'react';
 import { useMediatorAsState } from '../../hooks';
 import { FusionMediator } from '../../types';
 
 type MediatorProviderProps<
 	TData extends Record<PropertyKey, unknown>,
-	TContext extends Record<PropertyKey, unknown> = never
+	TContext extends Record<PropertyKey, unknown> = never,
+	TCustomSidesheetEvents extends BaseEvent = never
 > = {
 	children: ReactNode;
-	mediator: FusionMediator<TData, TContext>;
+	mediator: FusionMediator<TData, TContext, TCustomSidesheetEvents>;
 };
 export const MediatorProvider = <
 	TData extends Record<PropertyKey, unknown>,
 	TError,
-	TContext extends Record<PropertyKey, unknown> = never
+	TContext extends Record<PropertyKey, unknown> = never,
+	TCustomSidesheetEvents extends BaseEvent = never
 >({
 	children,
 	mediator,
-}: MediatorProviderProps<TData, TContext>) => (
+}: MediatorProviderProps<TData, TContext, TCustomSidesheetEvents>) => (
 	<MediatorContext.Provider
 		value={useMediatorAsState(mediator as unknown as FusionMediator<Record<PropertyKey, unknown>, never>)}
 	>