+++ packages/workspace-fusion/src/lib/hooks/useFilteredData.ts
@@ -1,10 +1,12 @@
+import { BaseEvent } from '@equinor/workspace-core';
 import { useState, useEffect } from 'react';
 import { FusionMediator } from '../types/fusionController';
 
 export function useFilteredData<
 	TData extends Record<PropertyKey, unknown>,
-	TContext extends Record<PropertyKey, unknown> = never
->({ dataService }: FusionMediator<TData, TContext>) {
+	TContext extends Record<PropertyKey, unknown> = never,
+	TCustomSidesheetEvents extends BaseEvent = never
+>({ dataService }: FusionMediator<TData, TContext, TCustomSidesheetEvents>) {
 	const [data, setData] = useState<TData[] | undefined>(dataService.filteredData);
 
 	useEffect(() => {