+++ packages/workspace-fusion/src/lib/integrations/garden/components/wrapper/GardenWrapper.tsx
@@ -1,3 +1,4 @@
+import { BaseEvent } from '@equinor/workspace-core';
 import { GardenController, Garden } from '@equinor/workspace-garden';
 import { NoDataSplashScreen } from '../../../../components/NoDataSplashScreen';
 import { FusionMediator } from '../../../../types';
@@ -7,10 +8,11 @@ type GardenWrapperProps<
 	TExtendedGardenFields extends string = never,
 	TCustomGroupByKeys extends Record<PropertyKey, unknown> = never,
 	TError extends Record<PropertyKey, unknown> = Record<PropertyKey, unknown>,
-	TContext extends Record<PropertyKey, unknown> = never
+	TContext extends Record<PropertyKey, unknown> = never,
+	TCustomSidesheetEvents extends BaseEvent = never
 > = {
 	controller: GardenController<TData, TExtendedGardenFields, TCustomGroupByKeys, TContext>;
-	mediator: FusionMediator<TData, TContext>;
+	mediator: FusionMediator<TData, TContext, TCustomSidesheetEvents>;
 };
 
 export const GardenWrapper = <
@@ -18,11 +20,12 @@ export const GardenWrapper = <
 	TExtendedGardenFields extends string = never,
 	TCustomGroupByKeys extends Record<PropertyKey, unknown> = never,
 	TError extends Record<PropertyKey, unknown> = Record<PropertyKey, unknown>,
-	TContext extends Record<PropertyKey, unknown> = never
+	TContext extends Record<PropertyKey, unknown> = never,
+	TCustomSidesheetEvents extends BaseEvent = never
 >({
 	controller,
 	mediator,
-}: GardenWrapperProps<TData, TExtendedGardenFields, TCustomGroupByKeys, TError, TContext>) => {
+}: GardenWrapperProps<TData, TExtendedGardenFields, TCustomGroupByKeys, TError, TContext, TCustomSidesheetEvents>) => {
 	return (
 		<NoDataSplashScreen mediator={mediator}>
 			<Garden controller={controller} />