+++ packages/workspace-fusion/src/lib/integrations/grid/utils/configureDataChange.ts
@@ -1,11 +1,13 @@
 import { GridController } from '@equinor/workspace-ag-grid';
+import { BaseEvent } from '@equinor/workspace-core';
 import { FusionMediator } from '../../../types';
 
 /**Update data on gridController whenever filtered data on mediator changes */
 export function configureDataChange<
 	TData extends Record<PropertyKey, unknown>,
-	TContext extends Record<PropertyKey, unknown> = never
->(gridController: GridController<TData, TContext>, mediator: FusionMediator<TData, TContext>) {
+	TContext extends Record<PropertyKey, unknown> = never,
+	TCustomSidesheetEvents extends BaseEvent = never
+>(gridController: GridController<TData, TContext>, mediator: FusionMediator<TData, TContext, TCustomSidesheetEvents>) {
 	mediator.dataService.filteredData$.subscribe((data) => {
 		if (!data) return;
 		gridController.rowData = data;