+++ packages/power-bi/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-powerbi@0.1.2...@equinor/workspace-powerbi@0.1.3) (2022-12-07)
+
+**Note:** Version bump only for package @equinor/workspace-powerbi
+
 ## [0.1.2](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-powerbi@0.1.1...@equinor/workspace-powerbi@0.1.2) (2022-11-30)
 
 ### Bug Fixes