+++ packages/status-bar/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.10](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-status-bar@0.0.9...@equinor/workspace-status-bar@0.0.10) (2022-12-07)
+
+**Note:** Version bump only for package @equinor/workspace-status-bar
+
 ## [0.0.9](https://github.com/equinor/fusion-workspace/compare/@equinor/workspace-status-bar@0.0.8...@equinor/workspace-status-bar@0.0.9) (2022-11-30)
 
 **Note:** Version bump only for package @equinor/workspace-status-bar